add pre-trained models and PyTorch version to download script
The GB-Net authors provides a collection of pre-trained models that can
be used to replicate the results from the paper. Furthermore, the model
depends on a very specific version of PyTorch that has to be manually
downloaded. The download script, Conda environment file, and README have
been updated to reflect this.